requirements.txt
django==1.8.17
django-tastypie
